Split keygen and signing tests into separate files
